cassandra_part/requirements.txt
kafka-python==2.0.2
cassandra-driver==3.25.0
flask==2.1.2
python-dateutil==2.8.2